test00 good up until the call to quit
Switched up somethings with when processes are started to match the
spec. I think the main issue was manually checking the priorities and
doing a context switch is fork1. In phase1a we don't need to worry about
that yet, and in phase1b we'll need to create a dispatcher to do that.
But also to check the priorities after the fork, we would need to switch the > to a < when comparing the
priorities, since a lower priority number means a greater priority. That
explains why the code was breaking even though the idea was right.